+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#1769](https://github.com/bbatsov/rubocop/issues/1769): Fix bug where `LiteralInInterpolation` registers an offense for interpolation of `__LINE__`. ([@rrosenblum][])
 * [#1773](https://github.com/bbatsov/rubocop/pull/1773): Fix typo ('strptime' -> 'strftime') in `Rails/TimeZone`. ([@palkan][])
 * [#1777](https://github.com/bbatsov/rubocop/pull/1777): Fix offense message from Rails/TimeZone. ([@mzp][])
+* Fix handling of `while` and `until` with assignment in `IndentationWidth`. ([@lumeet][])
 
 ## 0.30.0 (06/04/2015)